test: split register test coverage modules
